pyproject.toml
[project]
name = "service-bus"
version = "0.1.0"
description = "Learning and Development: Azure Service Bus"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "azure-servicebus>=7.13.0",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "setuptools>=75.7.0",
]
